+++ packages/opds-web-client/src/components/__tests__/OPDSCatalog-test.tsx
@@ -4,11 +4,11 @@ import { stub } from "sinon";
 import * as React from "react";
 import * as PropTypes from "prop-types";
 import { shallow } from "enzyme";
+import { ReactReduxContext } from "react-redux";
 
 import OPDSCatalog from "../OPDSCatalog";
 import Root, { RootProps } from "../Root";
-import buildStore from "../../store";
-import { mockRouterContext } from "./routing";
+import { mockRouterContext } from "../../__mocks__/routing";
 
 describe("OPDSCatalog", () => {
   let props = {
@@ -29,32 +29,6 @@ describe("OPDSCatalog", () => {
   };
   let context = mockRouterContext();
 
-  it("creates a store for Root if not given one", () => {
-    let wrapper = shallow(<OPDSCatalog {...props} />, {
-      context,
-      childContextTypes: {
-        router: PropTypes.object,
-        pathFor: PropTypes.func
-      }
-    });
-    let root = wrapper.find<RootProps>(Root);
-    expect(root.props().store).to.be.ok;
-  });
-
-  it("passes state to Root if given one", () => {
-    let store = buildStore();
-    let state = store.getState();
-    let wrapper = shallow(<OPDSCatalog {...props} initialState={state} />, {
-      context,
-      childContextTypes: {
-        router: PropTypes.object,
-        pathFor: PropTypes.func
-      }
-    });
-    let root = wrapper.find<RootProps>(Root);
-    expect(root.props().store.getState()).to.deep.equal(state);
-  });
-
   it("passes props to Root", () => {
     let wrapper = shallow(<OPDSCatalog {...props} />, {
       context,
@@ -63,8 +37,17 @@ describe("OPDSCatalog", () => {
         pathFor: PropTypes.func
       }
     });
-    let root = wrapper.find<RootProps>(Root);
+    /**
+     * This is painfully fragile, but must be done to make
+     * enzyme render the child function beneath ReactReduxContext.Consumer
+     */
+    const root = wrapper
+      .dive()
+      .dive()
+      .dive()
+      .dive();
 
+    // test that all of the props we passed in are present there
     Object.keys(props).forEach(key => {
       expect(root.props()[key]).to.equal(props[key]);
     });