+++ packages/opds-web-client/src/components/context/PathForContext.tsx
@@ -1,0 +1,35 @@
+import * as React from "react";
+import { PathFor } from "../../interfaces";
+import * as PropTypes from "prop-types";
+
+/**
+ * This is a component that will pass the pathFor prop down the tree
+ * via both old and new context apis.
+ */
+
+export const PathForContext = React.createContext<PathFor>(null);
+
+type PathForProps = {
+  pathFor: PathFor;
+  children: React.ReactChild;
+};
+
+export default class PathForProvider extends React.Component<PathForProps> {
+  static childContextTypes: React.ValidationMap<{}> = {
+    pathFor: PropTypes.func.isRequired
+  };
+
+  getChildContext() {
+    return {
+      pathFor: this.props.pathFor
+    };
+  }
+
+  render() {
+    return (
+      <PathForContext.Provider value={this.props.pathFor}>
+        {this.props.children}
+      </PathForContext.Provider>
+    );
+  }
+}